+++ src/Animate/Animate.meta.js
@@ -1,0 +1,8 @@
+import Animate from './Animate';
+import Registry from '@ui-autotools/registry';
+
+const metadata = Registry.getComponentMetadata(Animate);
+
+metadata.exportedFrom({
+  path: 'src/Animate/Animate.js',
+});