+++ src/Dropdown/Dropdown.js
@@ -1,12 +1,10 @@
-import classNames from 'classnames';
 import defaultTo from 'lodash/defaultTo';
 import isEqual from 'lodash/isEqual';
 import sortBy from 'lodash/sortBy';
 import { allValidators, extendPropTypes } from '../utils/propTypes';
-
 import InputWithOptions from '../InputWithOptions';
 import DropdownLayout from '../DropdownLayout';
-import styles from './Dropdown.scss';
+import styles from './Dropdown.st.css';
 
 const NO_SELECTED_ID = null;
 
@@ -89,9 +87,8 @@ class Dropdown extends InputWithOptions {
   }
 
   inputClasses() {
-    return classNames(styles.showPointer, {
-      [styles.noBorder]: this.props.noBorder,
-    });
+    const { noBorder } = this.props;
+    return styles('showPointer', { noBorder });
   }
 
   dropdownAdditionalProps() {