+++ src/index.js
@@ -170,4 +170,5 @@ export { default as HorizontalTimeline } from './HorizontalTimeline';
 export { default as StarsRatingBar } from './StarsRatingBar';
 export { default as TagList } from './TagList';
 export { default as FileUpload } from './FileUpload';
+export { default as Animate } from './Animate';
 export { default as MarketingPageTestimonialsFooter } from './MarketingPageTestimonialsFooter';