+++ stories/index.js
@@ -10,6 +10,7 @@ require('./Playground/Playground.story');
 // Components API
 require('../src/Accordion/docs/index.story');
 require('../src/AddItem/docs/index.story');
+require('../src/Animate/docs/index.story');
 require('../src/AnnouncementModalLayout/docs/index.story');
 require('../src/AudioPlayer/docs/index.story');
 require('../src/AutoComplete/docs/index.story');